app/requirements.txt
wheel==0.34.2
Flask==1.1.2
Werkzeug==1.0.1
python-libnmap==0.7.0
Flask-FontAwesome==0.1.4